T-2557: normalize table for logs as well.
