+++ packages/ui/src/components/Table/index.ts
@@ -1,0 +1,3 @@
+import Table from "./Table"
+export { Table }
+export default Table
